Script to download all files from a given website
Script to download all files of a given extension from a website. Bit of
overkill but wanted to play around with regex's a bit and url
verification although that turned out to be a serious rabbit hole.
Script works fine now but should do something about http:// in url and
make chunk size a parameter. Break still in for loop for debugging.